src/requirements.txt
pandas
google-cloud-bigquery
requests
